+++ Napping_PJ/Napping_PJ/Models/Entity/HotelExtraService.cs
@@ -7,7 +7,7 @@ namespace Napping_PJ.Models.Entity
     {
         public int HotelId { get; set; }
         public int ExtraServiceId { get; set; }
-        public int ExtraServicePrice { get; set; }
+        public double ExtraServicePrice { get; set; }
 
         public virtual ExtraService ExtraService { get; set; } = null!;
         public virtual Hotel Hotel { get; set; } = null!;
